backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
mysql-connector-python==8.0.26
requests==2.26.0
gunicorn==20.1.0
python-dateutil==2.8.2 